refactor code and reuse shared test helpers for deprovision tests
