+++ tests/lib/cdn/deprovision.py
@@ -62,17 +62,6 @@ def subtest_deprovision_removes_ALIAS_records(tasks, route53):
     route53.assert_no_pending_responses()
 
 
-def subtest_deprovision_removes_TXT_records(tasks, route53):
-    route53.expect_remove_TXT(
-        "_acme-challenge.example.com.domains.cloud.test", "example txt"
-    )
-    route53.expect_remove_TXT("_acme-challenge.foo.com.domains.cloud.test", "foo txt")
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    route53.assert_no_pending_responses()
-
-
 def subtest_deprovision_disables_cloudfront_distribution(
     instance_model, tasks, service_instance, cloudfront
 ):
@@ -180,46 +169,3 @@ def subtest_deprovision_removes_cloudfront_distribution_when_missing(
     )
     tasks.run_queued_tasks_and_enqueue_dependents()
     cloudfront.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_certificate_from_iam(
-    instance_model, tasks, service_instance, iam_commercial
-):
-    service_instance = db.session.get(instance_model, "1234")
-    iam_commercial.expects_delete_server_certificate(
-        service_instance.new_certificate.iam_server_certificate_name
-    )
-    iam_commercial.expects_delete_server_certificate(
-        service_instance.current_certificate.iam_server_certificate_name
-    )
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    iam_commercial.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_certificate_from_iam_when_missing(
-    instance_model, tasks, service_instance, iam_commercial
-):
-    service_instance = db.session.get(instance_model, "1234")
-    iam_commercial.expects_delete_server_certificate_returning_no_such_entity(
-        name=service_instance.new_certificate.iam_server_certificate_name
-    )
-    iam_commercial.expects_delete_server_certificate_returning_no_such_entity(
-        name=service_instance.current_certificate.iam_server_certificate_name
-    )
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    iam_commercial.assert_no_pending_responses()
-
-
-def subtest_deprovision_marks_operation_as_succeeded(instance_model, tasks):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "1234")
-    assert not service_instance.deactivated_at
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "1234")
-    assert service_instance.deactivated_at
-
-    operation = service_instance.operations.first()
-    assert operation.state == "succeeded"
